Fix building on the x86-64 architecture
In x86-64 unsigned long long is not considered the same type as
uint64_t so it was complaining that the declaration for
uprof_get_system_counter did not match the definition. It was also
throwing warnings where %llu was used to format a 64-bit value.

As glib is used throughout the program, it might as well use guint64
to get a 64-bit type and then we can use G_GUINT64_FORMAT to get the
format specifier for it.

+++ uprof/uprof-private.h
@@ -51,24 +51,24 @@ typedef struct _UProfCounterState
 typedef struct _UProfTimerState UProfTimerState;
 struct _UProfTimerState
 {
-  UProfObjectState    object;
+  UProfObjectState  object;
 
-  gboolean            disabled;
-  int                 recursion;
+  gboolean          disabled;
+  int               recursion;
 
-  char               *parent_name;
+  char             *parent_name;
 
-  unsigned long       count;
-  unsigned long long  start;
-  unsigned long long  total;
-  unsigned long long  partial_duration;
+  unsigned long     count;
+  guint64           start;
+  guint64           total;
+  guint64           partial_duration;
 
-  unsigned long long  fastest;
-  unsigned long long  slowest;
+  guint64           fastest;
+  guint64           slowest;
 
   /* note: not resolved until sorting @ report time */
-  UProfTimerState    *parent;
-  GList              *children;
+  UProfTimerState  *parent;
+  GList            *children;
 
 };
